pyproject.toml
[tool.poetry]
name = "abstractbox"
version = "0.1.0"
description = "Experimental container orchestrator"
authors = ["D0rs4n", "HyperGH"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pydantic = ">=1.9.2"
fastapi = ">=0.79.0"
rich = "^12.5.1"
typer = "^0.6.1"
PyJWT = "^2.4.0"
docker = "^6.0.0"
uvicorn = "^0.18.3"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.6"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length     = 120
target-version  = ["py310"]
include         = ".*py$"

[tool.isort]
profile = "black"
force_single_line = true

[tool.mypy]
strict = true

check_untyped_defs = true
incremental = true
namespace_packages = true
no_implicit_optional = true
implicit_reexport = true
pretty = true
python_version = "3.10"
show_column_numbers = true
show_error_codes = true
show_error_context = true

# stuff to allow
allow_untyped_globals = false
allow_redefinition = true

# stuff to disallow
disallow_untyped_decorators = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

# warnings
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

# dependency management (install stubs or fallback to Any)
install_types = true
non_interactive = true
ignore_missing_imports = true
